fix: Rename xcassets files to avoid collisions
Unfortunately, Xcode gets upset if it finds multiple ‘Assets.xcassets’ files in a Swift package source tree, even though the they’re not included in the package and the Swift Package Manager itself is happy. This change renames the assets to use non-standard names to reduce the risk of these conflicts.